<refactor>[front] refactor post fetching component
